Improving
Added some character class and attribute reference data files.
Its coming together.
Added a hack method to get first character to add all points to.
Need to:
Write a method to get the character name and display that with the
points.
Write method to display first name when logged in.
Duplicate signup page so it takes Jason to the create character page.